Unify water surfaces and enlarge watering can opening
